Initial design

See merge request greenaddress/green_ios!2

+++ gdk/tools/clean.sh
@@ -1,10 +1,17 @@
 #!/usr/bin/env bash
 set -e
 
-rm -fr build-*
-rm -fr subprojects/boost_1_66_0
-rm -fr subprojects/libwally-core-987575025520d18bac31e6e2d27c8c936d812c64
-rm -fr subprojects/autobahn-cpp-d0fcb61cb661241e0c2802abce85df359eb609fd
-rm -fr subprojects/websocketpp-294cc0d42957078176ff8f39397862e089838543
-rm -fr subprojects/openssl-OpenSSL_1_0_2o
-rm -fr subprojects/packagecache
+clean_meson() {
+  find subprojects/ -mindepth 2 -maxdepth 2 -name '*meson*' -not -name '*build*' | xargs rm -rf --
+}
+
+if [ "$1" == "meson" ]; then
+  clean_meson
+else
+  rm -fr build-*
+  clean_meson
+  find subprojects/ -maxdepth 1 -mindepth 1 -not -name '*meson*' -not -name '*wrap*' | xargs rm -rf --
+fi
+
+rm -rf docs/build docs/source/session.rst
+rm -rf .venv
